+++ src/main/java/com/project/shop/controller/items/GameController.java
@@ -1,10 +1,12 @@
-package com.project.shop.controller.item;
+package com.project.shop.controller.items;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
+@RequiredArgsConstructor
 @RequestMapping("/items")
 public class GameController {
