+++ src/main/java/com/project/shop/entity/Review.java
@@ -1,0 +1,27 @@
+package com.project.shop.entity;
+
+import jakarta.persistence.*;
+import org.springframework.data.annotation.CreatedDate;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name = "review")
+public class Review {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "review_id")
+    private Long id;
+
+    @Column(name = "review_title")
+    private String title;
+
+    @Column(name = "review_content")
+    private String content;
+
+    @Column(name = "review_date")
+    @CreatedDate
+    private LocalDateTime date;
+}
+
