+++ src/main/resources/templates/items/game/root.html
@@ -7,20 +7,24 @@
 	<link th:href="@{/css/gameRoot.css}" rel="stylesheet" />
 </th:block>
 
+<!-- 메인 틀-->
 <div id="mainFrame" layout:fragment="content">
+	<!-- 상단 카테고리 영역 -->
     <article class="cateList">
         <div class="categoryName">
             카테고리 대분류
         </div>
+		<!-- 검색 박스 -->
         <div class="searchBox">
             <input type="text">
             <button type="submit">검색or아이콘</button>
         </div>
     </article>
 
+	<!-- 메인 상품 표시 구역 -->
     <article class="mainList">
 		<ul class="mItemList">
-			<!-- 표시할 아이템 리스트. 한 줄에 4개 * 3줄 -->
+			<!-- 표시할 아이템 리스트. 한 줄에 4개 * 3줄 (예시) -->
 			<li class="itemBox1">
 				<div class="thumbnail">
 				</div>
