+++ src/main/resources/templates/notice/noticeReadArticle.html
@@ -8,7 +8,9 @@
 </th:block>
 
 <div id="mainFrame" layout:fragment="content">
+	<!-- 전체 레이아웃 -->
 	<article class="mainWindow">
+		<!-- 글쓰기 레이아웃 -->
 		<div class="window">
 			<div class="titleWindow">
 				제목
@@ -20,6 +22,7 @@
 			<div class="bodyWindow">
 				내용
 			</div>
+			<!-- 수정, 삭제를 위한 버튼 -->
 			<div class="modifyButton">
 				<div>
 					수정
