+++ src/main/resources/templates/support/faq.html
@@ -8,10 +8,13 @@
 </th:block>
 <div id="mainFrame" layout:fragment="content">
     <article class="mainWindow">
+        <!-- 상단 페이지 제목 -->
         <div class="mainTitle">
             <p>자주 묻는 질문</p>
+            <!-- 구분선 -->
             <hr>
         </div>
+        <!-- 부트스트랩 아코디언(펼치기-접기) 추후 scss 스타일 활용하여 색상 변경할 것 -->
         <div class="accordion" id="accordionPanelsStayOpenExample">
             <div class="accordion-item">
                 <h2 class="accordion-header" id="panelsStayOpen-headingOne">
